% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMGP.base.R
\name{CMGP.base}
\alias{CMGP.base}
\title{GBLUP model for genomic prediction}
\usage{
CMGP.base(
  y = NULL,
  CV = NULL,
  geno = NULL,
  random = NULL,
  EMsteps = 0L,
  EMsteps_fail = 10L,
  EM_alpha = 1,
  max_iter = 50L,
  eps = 0.01,
  doTrans = FALSE,
  get.KR = FALSE,
  get.P = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{random}{list, non-additive genetic random effect}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}

\item{doTrans}{logical, should marker effects be calculated}

\item{get.KR}{logical, should the Kinship be returned}

\item{get.P}{see ?gaston::lmm.aireml}

\item{verbose}{logical, TRUE for print the output}
}
\value{
list
}
\description{
GBLUP model for genomic prediction
}
\examples{
#\donttest{
#revl <- CMGP.base(y = y, CV = CV, geno = geno, random = random,
#EMsteps = EMsteps, EMsteps_fail = EMsteps_fail,
#EM_alpha = EM_alpha, max_iter = max_iter,
#eps = eps, doTrans = doTrans, get.KR = get.KR,
#get.P = get.P, verbose = verbose)
#}
}
