% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMGP.GF.R
\name{CMGP.GF}
\alias{CMGP.GF}
\title{Genomin factor BLUP}
\usage{
CMGP.GF(
  y = NULL,
  CV = NULL,
  geno = NULL,
  map = NULL,
  random = NULL,
  alpha = c(-3, 1),
  ldscore = NULL,
  bin = 1e+06,
  ncpus = 1,
  part.method = "ebvTrans",
  part.threshold = 0.01,
  LD.threshold = 0.7,
  doTrans = FALSE,
  get.KR = FALSE,
  get.P = FALSE,
  EMsteps = 0L,
  EMsteps_fail = 10L,
  EM_alpha = 1,
  max_iter = 50L,
  eps = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{map}{geno information, SNP, Chr, Pos}

\item{random}{list, non-additive genetic random effect}

\item{alpha}{genetype scaling value}

\item{ldscore}{LD score}

\item{bin}{region size for correlation calculation}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{part.method}{prioritization method}

\item{part.threshold}{significant level}

\item{LD.threshold}{cluster threshold}

\item{doTrans}{logical, TRUE for print the output}

\item{get.KR}{logical, should the Kinship be returned}

\item{get.P}{see ?gaston::lmm.aireml}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}

\item{verbose}{logical, TRUE for print the output}
}
\value{
list
}
\description{
Genomin factor BLUP
}
\examples{
#\donttest{
#revl <- CMGP.GF(y = y, CV = CV, geno = geno, map = map, random = random,
#alpha = alpha, ldscore = ldscore, bin = bin, ncpus = ncpus,
#part.method = part.method, part.threshold = part.threshold,
#LD.threshold = LD.threshold, doTrans = doTrans,
#get.KR = get.KR, get.P = get.P, EM_alpha = EM_alpha,
#EMsteps = EMsteps, EMsteps_fail = EMsteps_fail,
#max_iter = max_iter, eps = eps, verbose = verbose)
#}
}
