% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMGP.R
\name{CMGP}
\alias{CMGP}
\title{A Comprehensive Mixed-linear-model-based Tool for Genomic Prediction}
\usage{
CMGP(
  y = NULL,
  CV = NULL,
  geno = NULL,
  map = NULL,
  random = NULL,
  weight = NULL,
  alpha = c(-3, 1),
  ldscore = NULL,
  bin = 1e+06,
  AMB.region = 1e+06,
  LRT.threshold = 0.01,
  model = c("Base", "GF", "MA", "KW", "AMB", "MAPS"),
  part.method = c("OLS", "ebvTrans", "h2"),
  interval_S = c(1, 15),
  part.threshold = 0.01,
  LD.threshold = 0.7,
  doTrans = FALSE,
  get.KR = FALSE,
  get.P = FALSE,
  EMsteps = 0L,
  eps = 0.01,
  EMsteps_fail = 10L,
  EM_alpha = 1,
  max_iter = 50L,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{y}{phenotype, n*1 vector, NAs refer to validation}

\item{CV}{covariates, n*q matrix, numeric format for quantitative, character or factor for discrete}

\item{geno}{genotype, m * n matrix/bigmatrix}

\item{map}{marker information, SNP, Chr, Pos}

\item{random}{list, non-additive genetic effects}

\item{weight}{vector, weight of markers, using marker names to name it}

\item{alpha}{a numeric value for genotype scaling}

\item{ldscore}{vector, the sum of adjust squared correlation between a marker and markers within the region(in bin size)}

\item{bin}{region size for correlation calculation}

\item{AMB.region}{region size for genome dividing, only used when model was "AMB"}

\item{LRT.threshold}{significant level for likelihood ratio test, only used when model was "AMB}

\item{model}{c("Base", "GF", "MA", "KW", "AMB", "MAPS"), the first is default}

\item{part.method}{c("OLS", "ebvTrans", "h2"), the first is default}

\item{interval_S}{interval for search the optimal number of stratification, only used when model was "MAPS"}

\item{part.threshold}{significant level for statistic test, only used when model was "GF" or "MA}

\item{LD.threshold}{threshold for clustering when LD removing is working}

\item{doTrans}{logical, FALSE is default, if TRUE, gwas by linear transform will work}

\item{get.KR}{logical, should the Kinship be output}

\item{get.P}{see ?gaston::lmm.aireml}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{ncpus}{integer value for number of threads used}

\item{verbose}{logical, should the running progress be printed}
}
\value{
list
}
\description{
A Comprehensive Mixed-linear-model-based Tool for Genomic Prediction
}
\examples{
#\donttest{
#CMGP <- function(y = NULL, CV = NULL, geno = NULL, map = NULL, random = NULL,
#weight = NULL, alpha = c(-3, 1), ldscore = NULL, bin = 1e6,
#AMB.region = 1e6, LRT.threshold = 0.01,
#model = c("Base", "GF", "MA", "KW", "AMB", "MAPS"),
#part.method = c("OLS", "ebvTrans", "h2"), interval_S = c(1,15),
#part.threshold = 0.01, LD.threshold = 0.7, doTrans = FALSE,
#get.KR = FALSE, get.P = FALSE, EMsteps = 0L, eps = 1e-02,
#EMsteps_fail = 10L, EM_alpha = 1, max_iter = 50L,
#ncpus = 1, verbose = TRUE)
#}
}
