% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMGP.version.R
\name{CMGP.version}
\alias{CMGP.version}
\title{CMGP.version: Print the version of CMGP}
\usage{
CMGP.version(verbose = TRUE)
}
\arguments{
\item{verbose}{logical, whether print}
}
\description{
CMGP.version: Print the version of CMGP
}
\examples{
#CMGP.version(TRUE)
}
