% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatrix.R
\name{CovMatrix}
\alias{CovMatrix}
\title{Transform the covariates as model matrix}
\usage{
CovMatrix(CV = NULL, n = NULL)
}
\arguments{
\item{CV}{a matrix or data.frame for covariates}

\item{n}{integer, number of individuals}
}
\value{
a model matrix
}
\description{
The quantitative cov. were inputed as numeric format and remained unchanged,
the discrete cov. were transformed as one hot model matrix
}
\examples{
cv1 <- rnorm(100)
cv2 <- rep(c("m", "f"), 50)
n <- 100
CV <- CovMatrix(cbind(cv1,cv2), 100)
}
