% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBV.trans.R
\name{EBV.trans}
\alias{EBV.trans}
\title{Solve the mixed model using AIREML and calculate marker effects by linear transform}
\usage{
EBV.trans(
  y = NULL,
  CV = NULL,
  geno = NULL,
  weight = NULL,
  random = NULL,
  ncpus = 1,
  EMsteps = 0,
  EMsteps_fail = 10,
  EM_alpha = 1,
  max_iter = 50,
  eps = 0.01,
  get.KR = FALSE,
  get.P = FALSE,
  doTrans = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{weight}{numeric vector, the weight of markers, using marker names to name it}

\item{random}{list, non-additive genetic random effect}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}

\item{get.KR}{logical, should the Kinship be returned}

\item{get.P}{see ?gaston::lmm.aireml}

\item{doTrans}{logical, should marker effects be calculated}

\item{verbose}{logical, TRUE for print the output}
}
\value{
list, GP and GWAS
}
\description{
Solve the mixed model using AIREML and calculate marker effects by linear transform
}
\examples{
#\donttest{
#mix <- EBV.trans(y = y, CV = CV, geno = genolist, weight = pve,
#random = random, EMsteps = EMsteps, EM_alpha = EM_alpha,
#EMsteps_fail = EMsteps_fail, max_iter = max_iter,
#eps = eps, get.KR = FALSE, get.P = FALSE,
#doTrans = FALSE, verbose = TRUE)
#}
}
