% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_cut.R
\name{genome_cut}
\alias{genome_cut}
\title{Title}
\usage{
genome_cut(geno = NULL, map = NULL, bin = 1e+06, ncpus = 1, verbose = TRUE)
}
\arguments{
\item{geno}{genotype, m * n matrix/bigmatrix}

\item{map}{marker information, SNP, Chr, Pos}

\item{bin}{region size for genome dividing}

\item{ncpus}{integer value for number of threads used}

\item{verbose}{logical, should the running progress be printed}
}
\value{
divided genome list
}
\description{
Title
}
\examples{
#\donttest{
#genolist <- genome_cut(geno = geno, map = map, bin = bin,
#ncpus = ncpus, verbose = verbose)
#}
}
