% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinship.Van.R
\name{Kinship.Van}
\alias{Kinship.Van}
\title{Calculate the Kinship in VanRaden's method}
\usage{
Kinship.Van(geno = NULL, weight = NULL, ncut = 1, ncpus = 1)
}
\arguments{
\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{weight}{numeric vector, the weight of markers}

\item{ncut}{integer value, default as 1, how many block should the geno matrix be divided into?}

\item{ncpus}{integer value, default as 1, how many threads used}
}
\value{
genetic relationship matrix
}
\description{
Calculate the Kinship in VanRaden's method
}
\examples{
#\donttest{
#K <- Kinship.Van(geno=geno)
#}
}
