% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.remove.R
\name{LD.remove}
\alias{LD.remove}
\title{Select the most important marker in each cluster}
\usage{
LD.remove(
  index = NULL,
  geno = NULL,
  value = NULL,
  LD.threshold = 0.7,
  verbose = TRUE
)
}
\arguments{
\item{index}{index of markers}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{value}{improtance}

\item{LD.threshold}{cluster threshold}

\item{verbose}{logical, TRUE for print the output}
}
\value{
index
}
\description{
Select the most important marker in each cluster
}
\examples{
#\donttest{
#index <- LD.remove(index = index, geno = geno, value = value,
#LD.threshold = LD.threshold, verbose = verbose)
#}
}
