% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC.R
\name{LDSC}
\alias{LDSC}
\title{Parallel calculate the LD score}
\usage{
LDSC(geno = NULL, map = NULL, bin = 1e+06, ncpus = 1, verbose = TRUE)
}
\arguments{
\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{map}{geno information, SNP, Chr, Pos}

\item{bin}{region size for correlation calculation}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{verbose}{logical, TRUE for print the output}
}
\value{
vertor
}
\description{
Parallel calculate the LD score
}
\examples{
#\donttest{
#ldscore <- LDSC(geno = geno, map = map, bin = bin,
#ncpus = ncpus, verbose = verbose)
#}
}
