% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPS.partition.R
\name{MAPS.partition}
\alias{MAPS.partition}
\title{Sratification of markers based on PVE}
\usage{
MAPS.partition(
  y = NULL,
  CV = NULL,
  geno = NULL,
  map = NULL,
  random = NULL,
  alpha = 0,
  ldscore = NULL,
  bin = 1e+06,
  interval_S = c(1, 15),
  ncpus = 1,
  verbose = TRUE,
  EMsteps = 0L,
  EMsteps_fail = 10L,
  EM_alpha = 1,
  max_iter = 50L,
  eps = 0.01
)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{map}{geno information, SNP, Chr, Pos}

\item{random}{list, non-additive genetic random effect}

\item{alpha}{genetype scaling value}

\item{ldscore}{LD score}

\item{bin}{region size for correlation calculation}

\item{interval_S}{interval of number of stratification when optimizing}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{verbose}{logical, TRUE for print the output}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}
}
\value{
list
}
\description{
Sratification of markers based on PVE
}
\examples{
#\donttest{
#mp <- MAPS.partition(y = y, CV = CV, geno = geno, map = map, random = random,
#alpha = alpha, ldscore = ldscore, bin = bin,
#interval_S = interval_S, EMsteps = EMsteps, eps = eps,
#EMsteps_fail = EMsteps_fail, EM_alpha = EM_alpha,
#max_iter = max_iter, ncpus = ncpus, verbose = verbose)
#}
}
