% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.fast.R
\name{ols.fast}
\alias{ols.fast}
\title{Parallel OLS excluding the additional covariates}
\usage{
ols.fast(y = NULL, CV = NULL, geno = NULL, ncpus = 1, verbose = TRUE)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{verbose}{logical, TRUE for print the output}
}
\value{
matrix Beta, SE, Pvalue
}
\description{
Parallel OLS excluding the additional covariates
}
\examples{
#\donttest{
#revl <- ols.fast(y = y, CV = CV, geno = geno,
#ncpus = ncpus, verbose = verbose)
#}
}
