% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP.prioritized.R
\name{SNP.prioritized}
\alias{SNP.prioritized}
\title{Prioritization of marker using OLS, ebvTrans or PVE}
\usage{
SNP.prioritized(
  y = NULL,
  CV = NULL,
  geno = NULL,
  map = NULL,
  random = NULL,
  alpha = 0,
  ldscore = NULL,
  bin = 1e+06,
  method = c("OLS", "ebvTrans", "h2"),
  EMsteps = 0L,
  EMsteps_fail = 10L,
  EM_alpha = 1,
  max_iter = 50L,
  eps = 0.01,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{y}{phenotype vector, NAs refer to validation}

\item{CV}{covariates}

\item{geno}{Genotype in numeric format, pure 0, 1, 2 matrix; m * n, m is marker size, n is population size}

\item{map}{geno information, SNP, Chr, Pos}

\item{random}{list, non-additive genetic random effect}

\item{alpha}{genetype scaling value}

\item{ldscore}{LD score}

\item{bin}{region size for correlation calculation}

\item{method}{c("OLS", "ebvTrans", "h2"), "OLS" was default}

\item{EMsteps}{see ?gaston::lmm.aireml}

\item{EMsteps_fail}{see ?gaston::lmm.aireml}

\item{EM_alpha}{see ?gaston::lmm.aireml}

\item{max_iter}{see ?gaston::lmm.aireml}

\item{eps}{see ?gaston::lmm.aireml}

\item{ncpus}{integer value, default as 1, how many threads used}

\item{verbose}{logical, TRUE for print the output}
}
\value{
matrix
}
\description{
Prioritization of marker using OLS, ebvTrans or PVE
}
\examples{
#\donttest{
#sp <- SNP.prioritized(y = y, CV = CV, geno = geno, map = map,
#random = random, alpha = alpha, ldscore = ldscore,
#bin = bin, method = "h2", EMsteps = EMsteps,
#EMsteps_fail = EMsteps_fail, EM_alpha = EM_alpha,
#max_iter = max_iter, eps = eps,
#ncpus = ncpus, verbose = verbose)
#}
}
