% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.trans.R
\name{time.trans}
\alias{time.trans}
\title{Transform the time from second format}
\usage{
\method{time}{trans}(x)
}
\arguments{
\item{x}{numeric, elapsed time in second}
}
\value{
a character
}
\description{
Transform the time from second format
}
\examples{
#time.trans(100)
}
